requirements.txt
streamlit
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pypdf
python-docx
imgkit